refactor(viewer): major simplification of renderer use
